+++ src/top/Persistencia/IRepoMarca.cs
@@ -2,6 +2,6 @@ namespace Topardex.top.Persistencia;
 
 public interface IRepoMarca
 {
-    public ushort IdMarca { get; set; }
+    public int IdMarca { get; set; }
     public string Nombre { get; set; }
 }
