+++ src/top/ProductoPedido.cs
@@ -2,10 +2,10 @@ namespace Topardex;
 
 public class ProductoPedido
 {
-    public ushort IdProductoPedido { get; set; }
+    public int IdProductoPedido { get; set; }
     public Producto Producto { get; set; }
 
-    public ushort Cantidad { get; set; }
+    public int Cantidad { get; set; }
     public decimal PrecioUnitario { get; set; }
 
     public decimal Subtotal => Cantidad * PrecioUnitario;
